♻️ refactor: 디렉토리 구조 개선 및 FAISS 벡터 검색 파라미터 조정
- 폴더 및 파일명 명확화
- FAISS 검색 결과 문서 수 top_k=2로 제한
- 청킹 chunk_size=512로 조정하여 embedding 효율성 향상
