+++ app/models/embedding_model.py
